Crate smc

smc

This crate provides a simple implementation of notable (academic) commitment schemes, on top of the popular OpenSSL cryptography library.

This is mostly an exercise to learn Rust. I took inspiration from the Secure Multiparty Computation class (@Sapienza Univerity of Rome).

Supported commitment schemes

El Gamal

Pedersen

They can be used on top of groups where Discrete Logaritm problem is assumed to be hard: currently, the crate supports both Multiplicative Group Zp*, for a safe prime p, and Elliptic Curve.

Modules

commitment
A simple commitment scheme implementation.

group
A simple wrapper for groups where Discrete Log problem is considered to be hard.

utils
Utilities for secure random number generation. This is a simple wrapper of OpenSSL rand module.