Struct smc::group::MultiplicativeGroup

pub struct MultiplicativeGroup { /* fields omitted */ }

Represents the multiplicative group Zp*, where p is a safe prime.

A safe prime p is such that p = 2q + 1, for a prime p. Note that when p is a safe prime, the multiplicative group of numbers modulo p has a subgroup of large prime order.

Methods

impl MultiplicativeGroup

pub fn new(secpar: i32) -> Result<Self, ErrorStack>