Trait smc::group::DLogGroup

pub trait DLogGroup<E: Element> {
    fn get_generator(&self) -> &E;

    fn get_order(&mut self) -> BigNum;

    fn generate_random_element(&self) -> E;

    fn generate_random_exponent(&self) -> BigNum;

    fn exponentiate(&mut self, e1: &E, e2: &BigNum) -> E;

    fn multiply(&mut self, e1: &E, e2: &E) -> E;

    fn pow(&mut self, pow: &BigNum) -> E;

    fn eq(&mut self, e1: &E, e2: &E) -> bool;
}

This is the trait for groups where Discrete Log (DL) problem is considered to be hard.

DL problem: given a generator g of a finite group G and a random element h in G, find the integer x such that g^x = h. The two most common classes in cryptographic applications are the multiplicative group Zp* for a large prime p, and some Elliptic curve group.

For more details check here.

Required methods

fn get_generator(&self) -> &E

Retrieves the generator g for the group.

fn get_order(&mut self) -> BigNum

Returns the order of the group.

fn generate_random_element(&self) -> E

Generates a random element in the group.

fn generate_random_exponent(&self) -> BigNum

Generates a random exponent

fn exponentiate(&mut self, e1: &E, e2: &BigNum) -> E

Computes an exponentiation between two elements in the group.

fn multiply(&mut self, e1: &E, e2: &E) -> E

Computes the multiplication between two elements in the group.

fn pow(&mut self, pow: &BigNum) -> E

Computes the pow.

fn eq(&mut self, e1: &E, e2: &E) -> bool

Compares two elements in a group.

Loading content...